Handle missing dependencies for package setup/build
Currently, the package setup requires the build dependencies `numpy` and
`cython`to be already installed when trying to install `cyvcf2`. As an
example, `pip install cyvcf2` with either one of the dependencies
missing fails with an `ImportError`. The issue arises since setuptools
requires parsing `setup.py` once before any install occurs, in order to
detect the dependency structure.

This patch allows missing dependencies when parsing `setup.py`, and
separates a) installing missing dependencies and b) build extensions
which require these dependencies to be present.