Install setup_requires dependencies with pip
The setup.py script needs to import Cython and numpy, but these are not
available when the setup.py script is run. The setuptools solution to this
is to add a setup_requires parameter to setup(). That has some problems
because it bypasses pip and installs dependencies with easy_install.
That does not know anything about wheels and is therefore a lot slower,
in particular when installing numpy.
<https://pip.pypa.io/en/latest/reference/pip_install/> has some warnings
about setup_requires.

The approach here is to spawn a 'pip install' process as part of
setup.py, as proposed by Daniel Holth in
<https://mail.python.org/pipermail/distutils-sig/2014-May/024391.html>.
The code is at <https://bitbucket.org/dholth/setup-requires/src>.
It is definitely a hack, but it seems to work well.

More discussions about setup_requires:
https://www.mail-archive.com/distutils-sig@python.org/msg25121.html
https://github.com/pypa/pip/issues/1820

Closes #28.