release gil:
this and the previous changes release the gil around critical sections of iteration
over a VCF. Those are intializing an iterator and parsing the bcf_record, either
from a tabixy query or just straight iteration.

In my test, I can create a threadpool, each with it's own vcf object, do region queries
in parallel and get near perfect parallelism with 16 threads.

Because cyvcf2 objects are not serializable, python threads are the only way to parallelize
operations of a VCF.