allow getting sequence lengths via vcf.seqlens
closes #64

this also changes the behavior of seqnames to first check the header
and return all values in the header. if no values are present in the
header then it will check for a tabix index and use that.

previously, it would check the tabix index first. this had the problem
of returning only returning the seqnames of chromosomes taht contained
variants. the new version will return any seq that's present in then
header even if a particular chrom does not have any variants. this
ensures that there is a 1::1 correspondence between seqnames and
the newly added seqlens.