Conditionally Cythonize the sources (#143)
This commit introduces an environment variable CYTHONIZE to control
whether to use Cython to generate the C code.

For example, the following will call Cython to generate the C code and
create source code tarball and wheel, useful for code release:

    CYTHONIZE=1 python setup.py sdist bdist_wheel

This command requires Cython (and other requirements) is available. When
Cython is not found, it warns the user to run:

    pip install -r requirements.txt

Alternatively, one can install cyvcf2 from the releases (``.tar.gz`` or
``.whl``),  this installation should not require Cython since the
generated C code has been included or compiled. For example,

    pip install dist/cyvcf2.tar.gz

This command now also correctly installs numpy, click, and other
dependencies.