Allow Reader/Writer to be context managers
By adding __enter__ and __exit__ methods to the Reader and Writer
classes, they can now be used as context managers. Also added a test to
make sure the Reader context manager works correctly. Did not add a test
to do the same with Writer, because trying to write to a closed Writer
results in a segfault rather than raising an exception that can be
checked by pytest.